requirements.txt
prometheus_client
flask
python-json-logger
opentelemetry-distro
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-httpx
opentelemetry-exporter-otlp
mysql-connector-python
opentelemetry-instrumentation-mysql